+++ src/query/sql/src/planner/optimizer/optimizers/distributed/mod.rs
@@ -17,6 +17,6 @@ mod distributed;
 mod distributed_merge;
 mod sort_and_limit;
 
-pub use distributed::optimize_distributed_query;
+pub use distributed::DistributedOptimizer;
 pub use distributed_merge::BroadcastToShuffleOptimizer;
 pub use sort_and_limit::SortAndLimitPushDownOptimizer;
